+++ server/lib/providers/github.js
@@ -1,4 +1,5 @@
 import _ from 'lodash';
+import path from 'path';
 import axios from 'axios';
 import Promise from 'bluebird';
 import GitHubApi from 'github';
@@ -255,6 +256,17 @@ const getEmailProvider = (repository, branch, files) =>
     );
 
 /*
+ * Get tenant settings.
+ */
+const getTenant = (repository, branch, files) =>
+  downloadConfigurable(
+    repository,
+    branch,
+    'tenant',
+    { configFile: _.find(files, f => utils.isTenantFile(f.path)) }
+    );
+
+/*
  * Get all configurables (resource servers / clients).
  */
 const getConfigurables = (repository, branch, files, directory) => {
@@ -277,9 +289,13 @@ export const getChanges = ({ repository, branch, sha }) =>
 
       const promises = {
         rules: getRules(repository, branch, files),
+        tenant: getTenant(repository, branch, files),
         databases: getDatabaseData(repository, branch, files),
         emailProvider: getEmailProvider(repository, branch, files),
         emailTemplates: getHtmlTemplates(repository, branch, files, constants.EMAIL_TEMPLATES_DIRECTORY, constants.EMAIL_TEMPLATES_NAMES),
+        guardianFactors: getConfigurables(repository, branch, files, path.join(constants.GUARDIAN_DIRECTORY, constants.GUARDIAN_FACTORS_DIRECTORY)),
+        guardianFactorTemplates: getConfigurables(repository, branch, files, path.join(constants.GUARDIAN_DIRECTORY, constants.GUARDIAN_TEMPLATES_DIRECTORY)),
+        guardianFactorProviders: getConfigurables(repository, branch, files, path.join(constants.GUARDIAN_DIRECTORY, constants.GUARDIAN_PROVIDERS_DIRECTORY)),
         pages: getHtmlTemplates(repository, branch, files, constants.PAGES_DIRECTORY, constants.PAGE_NAMES),
         roles: getConfigurables(repository, branch, files, constants.ROLES_DIRECTORY),
         clients: getConfigurables(repository, branch, files, constants.CLIENTS_DIRECTORY),