+++ tests/server/mocks/repo-tree-mock.js
@@ -77,5 +77,40 @@ module.exports = {
   },
   pages: {
     'login.html': 'login page html'
+  },
+  'guardian/factors': {
+    'sms.json': {
+      name: 'sms',
+      enabled: true
+    },
+    'email.json': {
+      name: 'email',
+      enabled: false
+    }
+  },
+  'guardian/providers': {
+    'sms-twilio.json': {
+      name: 'sms',
+      provider: 'twilio',
+      auth_token: 'test_twilio_authtoken',
+      sid: 'test_twilio_sid',
+      from: '0800-TEST-NUMBER',
+      messaging_service_sid: 'test_copilot_sid'
+    }
+  },
+  'guardian/templates': {
+    'sms.json': {
+      name: 'sms',
+      enrollment_message: 'test enroll {{ code }}',
+      verification_message: 'test verification {{ code }}'
+    }
+  },
+  'tenant.json': {
+    friendly_name: 'My Company',
+    support_email: 'support@company.com',
+    session_lifetime: 1.23,
+    default_directory: 'users',
+    sandbox_version: '4',
+    idle_session_lifetime: 72
   }
 };